Conflicts:
	src/motion-module.cpp

+++ src/sync.h
@@ -44,7 +44,7 @@ namespace rsimpl
 
         const byte * get_frame_data(rs_stream stream) const;
         double get_frame_timestamp(rs_stream stream) const;
-        int get_frame_number(rs_stream stream) const;
+        unsigned long long get_frame_number(rs_stream stream) const;
         long long get_frame_system_time(rs_stream stream) const;
 
         frameset * clone_frontbuffer();
